73 Map Query bug: working now with findElec
Add state to candidates.json so we can draw and select from the right
layer after finding by elec name.
Now it works! Zooms to electorate. And then it zooms to Postcode as
well. Still draws elec lines and labels. Still zooms to elec when map
clicked.

Perhaps the workflow for findElec does something critical which was
missing from the workflow for findPc. If so, it will fail to zoom when
findPc is run first.